trackmania-rl-main/requirements.txt
spyder
spyder-kernels
isort
black
numpy
matplotlib
pydirectinput
joblib
mypy
pprofile
torchrl
dxcam
psutil
readwritememory
setuptools
typing
numba
opencv-python